Introduction

Introduction

Homotopy.jl is a package for constructing (polynomial) homotopies $H(x,t)$.

Each implemented homotopy has the same Interface so that you can switch easily between different homotopy types. Based on this interface there are also some convenient higher level constructs provided, e.g. the construction of a total degree system and its start solutions.

Example

using Homotopy
# we use an MultivariatePolynomials implementation to construct the homotopy.
import DynamicPolynomials: @polyvar

@polyvar x y z

H = StraightLineHomotopy([x + y^3, x^2*y-2y], [x^3+2, y^3+2])
# H is now StraightLineHomotopy{Int64},
# but let's assume our algorithm uses Complex128, to avoid unnecessary conversions
# it would be better to make
H = StraightLineHomotopy{Complex128}([x + y^3, x^2*y-2y], [x^3+2, y^3+2])

# we can now evaluate H
evaluate(H, rand(Complex128, 2), 0.42)
# or alternatively
H(rand(Complex128, 2), 0.42)

Homotopies

The following homotopies are implemented

Homotopy.StraightLineHomotopy — Type.

StraightLineHomotopy(start, target)

Construct the homotopy t * start + (1-t) * target.

start and target have to match and to be one of the following

Vector{<:MP.AbstractPolynomial} where MP is MultivariatePolynomials

MP.AbstractPolynomial

Vector{<:FP.Polynomial} where FP is FixedPolynomials

StraightLineHomotopy{T}(start, target)

You can also force a specific coefficient type T.

source

Homotopy.GammaTrickHomotopy — Type.

GammaTrickHomotopy(start, target, [γ])

Construct the homotopy (1-t)⋅target + t⋅γ⋅start. If γ is not supplied it will be drawn randomly and uniformly from the complex unit circle.

start and target have to match and to be one of the following

Vector{<:MP.AbstractPolynomial} where MP is MultivariatePolynomials

MP.AbstractPolynomial

Vector{<:FP.Polynomial} where FP is FixedPolynomials

GammaTrickHomotopy(start, target, seed::Int)

You can also supply a seed for the RNG which is used to draw γ, i.e. subsequent invocations with the same seed will produce the same γ.

GammaTrickHomotopy{T}(start, target, [γ])
GammaTrickHomotopy{T}(start, target, seed::Int)

You can also force a specific coefficient type T.

source