Higher level constructs

Higher level constructs

Homotopy.totaldegree — Function.

totaldegree(H::Type{AbstractHomotopy}, F, [unitroots=false])

Construct a total degree homotopy of type H with F and an iterator of its solutions. This is the homotopy with start system

\[\begin{align*} z_1^{d_1} &- b_1\\ z_1^{d_2} &- b_2\\ &\vdots \\ z_n^{d_n} &- b_n\\ \end{align*}\]

and target system F, where $d_i$ is the degree of the $i$-th polynomial of F. If unitroots=true then $b_i=1$ otherwise $b_i$ is a random complex number (with real and imaginary part in the unit interval).

Example

H, startsolutions = totaldegree(StraightLineHomotopy{Complex128}, [x^2+y+1, x^3*y-2])

source

Homotopy.TotalDegreeSolutionIterator — Type.

TotalDegreeSolutionIterator(degrees, b)

Given the Vectors degrees and b TotalDegreeSolutionIterator enumerates all solutions of the system

\[\begin{align*} z_1^{d_1} - b_1 &= 0 \\ z_1^{d_2} - b_2 &= 0 \\ &\vdots \\ z_n^{d_n} - b_n &= 0 \\ \end{align*}\]

where $d_i$ is degrees[i] and $b_i$ is b[i].

source

Homotopy.totaldegree_startsystem — Function.

totaldegree_startsystem(F::Vector{FP.Polynomial{<:Complex}}, [unit_roots=false])

Return the system

\[\begin{align*} z_1^{d_1} &- b_1\\ z_1^{d_2} &- b_2\\ &\vdots \\ z_n^{d_n} &- b_n\\ \end{align*}\]

where $d_i$ is the degree of the $i$-th polynomial of F and an iterator of its solutions. If unitroots=true then $b_i=1$ otherwise $b_i$ is a random complex number (with real and imaginary part in the unit interval).

source