Interface

Interface

Evaluation

Homotopy.evaluate — Function.

evaluate(H::AbstractHomotopy, x, t)

Evaluate the homotopy H at x to time t, i.e. $H(x,t)$

source

Homotopy.evaluate! — Function.

evaluate!(u::Vector, H::AbstractHomotopy, x, t)

Evaluate the homotopy H at x to time t, i.e. $H(x,t)$, and store the result in u. Use this instead of evaluate to avoid allocations.

source

Differentiation

Homotopy.jacobian — Function.

jacobian(H::AbstractHomotopy)

Compute an evaluation function (x, t) -> J_H(x,t) of the jacobian $J_H$ of the homotopy $H$. The jacobian is constructed w.r.t. x, i.e. it doesn't contain the partial derivatives w.r.t. t.

source

Homotopy.jacobian! — Function.

jacobian!(H::AbstractHomotopy)

Compute an inplace evaluation function (u, x, t) -> u := J_H(x,t) of the jacobian $J_H$ of the homotopy $H$. Use this instead of jacobian to avoid allocations.

source

Homotopy.dt — Function.

dt(H::AbstractHomotopy)

Compute an evaluation function (x, t) -> ∂H∂t(x,t) of the partial derivative $\frac{∂H}{∂t}$ of the homotopy $H$.

source

Homotopy.dt! — Function.

dt!(H::AbstractHomotopy)

Compute an inplace evaluation function (u, x, t) -> u := ∂H∂t(x,t) of the partial derivative $\frac{∂H}{∂t}$ of the homotopy $H$. Use this instead of dt to avoid allocations.

source

Homogenization

Homotopy.homogenize — Function.

homogenize(H::AbstractPolynomialHomotopy)

Homogenize the homotopy H. This adds an additional variable. If H is already homogenized, this is the identity.

source

Homotopy.dehomogenize — Function.

dehomogenize(H::AbstractPolynomialHomotopy)

Dehomogenize the homotopy H. This removes the first variable. If H is not homogenized, this is the identity.

source

Homotopy.ishomogenized — Function.

ishomogenized(H::AbstractPolynomialHomotopy)

Check whether the homotopy H was homogenized.

source

Homotopy.ishomogenous — Function.

ishomogenous(H::AbstractPolynomialHomotopy)

Check whether the homotopy H is homogenous. This does not imply that H was homogenized.

source

Misc

Homotopy.nvariables — Function.

nvariables(H::AbstractPolynomialHomotopy)

The number of variables which H expects as input, i.e. to evaluate H(x,t) x has to be a vector of length nvariables(H).

source

Homotopy.weylnorm — Function.

weylnorm(H::AbstractHomotopy)

Creates a function with variable t that computes the Weyl norm (or Bombieri norm) of $H(x,t)$. See here for details about the Weyl norm.

source